requirements.txt
torch
torchvision
numpy
scipy
scikit-learn
matplotlib
Pillow
PyYAML
ninja
